LayoutTests/http/tests/webgpu/webgpu/shader/validation/static_assert/static_assert.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:shader,validation,static_assert,static_assert:*";
                </script>
            
