LayoutTests/http/tests/webgpu/webgpu/shader/validation/static_assert/static_assert-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Query `webgpu:shader,validation,static_assert,static_assert:*` does not match any cases

Harness Error (FAIL), message = Unhandled rejection: Query `webgpu:shader,validation,static_assert,static_assert:*` does not match any cases
